==> code/Validate_Over_Concentrations/code/9_Analyse_Perturbations_Over_Concentrations.ipynb <==
# Analyse Perturbations over different Concentrations

## Libraries 

import pandas as pd
import numpy as np
from matplotlib import pylab as plt
from scipy import spatial
import os
from sklearn.metrics import cohen_kappa_score
import seaborn as sns
from scipy.stats import mannwhitneyu

## Functions and definitions

# Some Easy Outlier detection
def reject_outliers_2(data, m=6.):
    d = np.abs(data - np.median(data))
    mdev = np.median(d)
    s = d / (mdev if mdev else 1.)
    return [data[i] for i in range(0, len(data)) if s[i] < m]


def ensure_dir(file_path):
    '''
    Function to ensure a file path exists, else creates the path

    :param file_path:
    :return:
    '''
    directory = os.path.dirname(file_path)
    if not os.path.exists(directory):
        os.makedirs(directory)
        
        
########################################
##Actual Math for calculating the DDIs##
########################################


def unit_vector(vector):
    """ Returns the unit vector of the vector.  """
    return vector / np.linalg.norm(vector)
def angle_between(v1, v2):
    """ Returns the angle in radians between vectors 'v1' and 'v2'::

            angle_between((1, 0, 0), (0, 1, 0))
            1.5707963267948966
            angle_between((1, 0, 0), (1, 0, 0))
            0.0
            angle_between((1, 0, 0), (-1, 0, 0))
            3.141592653589793
    """
    v1_u = unit_vector(v1)
    v2_u = unit_vector(v2)
    return np.arccos(np.clip(np.dot(v1_u, v2_u), -1.0, 1.0))
def calculate_vector_math_v2(a, b, c):
    '''
    calculate the amount of single a, single b, and the 'surprise factor)
    :param a: vector a (single)
    :param b: vector b (single)
    :param c: vector c (combination)
    :return: alpha, beta and gamma (part of vector a, b and suprise)
    '''

    if sum(a) != 0 and sum(b) != 0:

        
        #Check first for two special cases:
        
        # 1.
        #Check if angle between the two vectors is to narrow (both point in same direction)
        if angle_between(a,b) <= 0.5:
            #h = c/(a+b)

            A=  np.array([[np.dot(a, a), np.dot(a, b)], [np.dot(a, b), np.dot(b, b)]])
            h = np.array([np.dot(a, c), np.dot(b, c)])

            alpha = h[0] / (A[0][0]+A[1][0])
            beta = h[1]  / (A[0][1]+A[1][1])

            n =alpha * a + beta * b - c

            gamma = np.linalg.norm(n)
            return str(alpha),str(beta),str(gamma)
        
        # 2. 
        #Check if angle between two is EXACTLY 180 degrees
        elif angle_between(a,b) == 3.141592653589793:
            A = np.array([[np.dot(a, a), np.dot(a, b)], [np.dot(a, b), np.dot(b, b)]])
            h = np.array([np.dot(a, c), np.dot(b, c)])


            alpha = h[0] / (A[0][0] + abs(A[1][0]))
            beta = h[1]  / (abs(A[0][1]) + A[1][1])

            n = alpha * a + beta * b - c

            gamma = np.linalg.norm(n)

            return str(alpha), str(beta), str(gamma)

    try:

        #
        # Check if all or any of the 3 vectors is zero
        
        if sum(c) != 0 and sum(a) == 0 and sum(b) == 0:
            return '1','1',str(dis.euclidean([0]*len(c),c))
        elif sum(c) == 0 and sum(a) == 0 and sum(b) == 0:
            return '1','1','0'
        elif sum(c) == 0 and sum(a) == 0:
            return '1','0','0'
        elif sum(c) == 0 and sum(b) == 0:
            return '0','1','0'

        else:
            
            
            # Matrix equation
            A = np.array([[np.dot(a, a), np.dot(a, b)], [np.dot(a, b), np.dot(b, b)]])

            h = np.array([np.dot(a, c), np.dot(b, c)])

            
            if A[0][0]==0 and h[0] ==0: #one vector zero, so combination can be only 1dim

                beta = h[1]/A[1][1]
                n = beta * b -c
                gamma = np.linalg.norm(n)

                if  (len(list(b)) - list(b).count(0) > 2) or np.linalg.norm(b) > 0.5:
                    return '1.0',str(beta),str(gamma)
                else:
                    return '1', '1', str(dis.euclidean([0] * len(c), c))
            elif A[1][1]==0 and h[1] ==0:
                alpha = h[0]/A[0][0]
                n = alpha * a -c
                gamma = np.linalg.norm(n)


                if len(list(a)) - list(a).count(0) > 2 or np.linalg.norm(a) > 0.5:
                    return str(alpha),'1',str(gamma)
                else:
                    return '1', '1', str(dis.euclidean([0] * len(c), c))
            elif h[0] == 0 and h[1] == 0:
                gamma = np.linalg.norm(c)
                return '0.0','0.0',str(gamma)
            elif A[0][0] != 0 and h[0] == 0 and h[1] == 0:
                gamma = np.linalg.norm(c)
                return '0.0','1.0',str(gamma)
            elif A[1][1] != 0 and h[0] == 0 and h[1] == 0:
                gamma = np.linalg.norm(c)
                return '1.0','0',str(gamma)

            
            ##########################
            # The ideal normal case with 3 vectors in correct angle and not being zero
            ##########################
            
            p = np.linalg.solve(A, h)
            # orthogonal vector
            n = p[0] * a + p[1] * b - c

            distance = np.linalg.norm(n)
            # check
            # print('dot product of a and c: %.4f' %(np.dot(a,n)))
            # print('dot product of b and c: %.4f' %(np.dot(b,n)))
            # print('distance: %.3f' %(distance))
            return str(p[0]), str(p[1]), str(distance)
    except:
        return 'Error', 'Error', 'Error'        


## Main functions

### Extract further metadata

per_well_CellCount = pd.read_csv('../results/IsabelCLOUPAC_Per_Image/CellCount/Individual_Well_Results.csv')
per_well_CellCount.Plate.astype(str)
per_well_CellCount.head()

list_of_clouds = []
fp = open('../results/IsabelCLOUPAC_Per_Image/List_of_CLOUD_Drugs/List_of_usable_IDs.csv','r')
for line in fp:
    list_of_clouds.append(line.strip())
fp.close()
print len(list_of_clouds)

perturbations = pd.read_csv('../results/IsabelCLOUPAC_Per_Image/PerturbationVectors/Vectors.csv')

features = perturbations.columns[1:].values

#Drug name contains: name, concentration, well and plate information
perturbations['Drug'] = perturbations['Perturbation'].str.split('|').str[0]
perturbations['Concentrations'] = perturbations['Perturbation'].str.split('|').str[1].str.split('_').str[0]
perturbations['Well'] = perturbations['Perturbation'].str.split('|').str[1].str.split('_').str[1]
perturbations['Plate'] = perturbations['Perturbation'].str.split('|').str[1].str.split('_').str[2]

perturbations.head()

perturbations["Plate"] = pd.to_numeric(perturbations["Plate"])
per_well_CellCount["Plate"] = pd.to_numeric(per_well_CellCount["Plate"])


perturbations = pd.merge(perturbations, per_well_CellCount,  how='left', left_on=['Well','Plate'], right_on = ['Well','Plate'])
perturbations.head(10)

# INCLUDE ONLY CLOUDS
### 
perturbations = perturbations.loc[perturbations['Drug'].isin(list_of_clouds)]
print len(perturbations)

# INCLUDE Wells with more than 50 cells
### 
perturbations = perturbations.loc[perturbations['CellCount'] > 25]
print len(perturbations)

### Calculate vector norms

#CLOUD139 had a transfer problem (hence excluded form the very begining)


found_CLOUDS = perturbations['Drug'].unique()
for cloud in list_of_clouds:
    #if len(perturbations.loc[perturbations['Drug'] ==cloud]) < 8:
    #    print 'CLOUD with not all replicates: %s' %cloud
    if cloud not in found_CLOUDS:
        print 'Not found: %s' %cloud

perturbations.head(10)

# CALCULATE the actual vector norm and add to the pandas dataframe
##

perturbations['VectorNorm'] = np.linalg.norm(perturbations[features].values,axis=1)

### Find significant perturbations
- Use the mean over all vector norms + 0.25 standard deviations as a mild threshold for including only significant drug perturbations

# Calculate threshold
##
threshold_significant = perturbations['VectorNorm'].mean() + 0.5 *perturbations['VectorNorm'].std()
#threshold_significant = 0.8

# Create a histogram showing all drugs being larger than the threshold
###

plt.hist(perturbations['VectorNorm'].values,bins='auto', color='#3AB9D1')
plt.axvline(threshold_significant, ls='--', color='red')
plt.xlabel('Vector Norm')
plt.ylabel('Frequency')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Significant_Perturbations.pdf')
plt.close()

# Show Output
##
print 'Number significant perturbations: %d' %len(perturbations.loc[perturbations['VectorNorm'] > threshold_significant])
print 'Number non significant perturbations: %d' %len(perturbations.loc[perturbations['VectorNorm'] <= threshold_significant])


print 'Significance Threshold: %.2f' %threshold_significant
print 'Number significant Perturbations: %.2f' %threshold_significant

# Number of valid drugs
##
all_drugs = perturbations['Drug'].unique()
print 'Number of drugs: %d' %len(all_drugs)

## Check if result for same drug concentrations is more similar than what expected from random
The idea is that if the concentration of the same drug is just increased, the effect should just change linear. Meaning that the two vectors should still both point in the same direction (cosine similarity/ alpha/beta) but with an incresed effect (delta vector norm)

- Check the Cosine similarity (measure of similarity/linearity)
- Delta vector norm  (measure of increase of effect size)
- alpha/beta contribution (another measure of linearity given by the novel mathematical framework)

### Calculate the real results

## Reald draws
####


# save results in vectors (consecutive concentrations)
all_AlphaBetas = [] #alpha/beta
all_Gammas = [] #gamma
all_cosine_distances = [] #cosine simialrity (actually similarity --> see below)
all_delta_VectorNorms = [] #Dehltave vector norm

### additional test variable
all_gamma_explained = []

# save results in vectors (consecutive concentrations
all_AlphaBetas_replicates = [] #alpha/beta
all_Gammas_replicates = [] #gamma
all_cosine_distances_replicates = [] #cosine simialrity (actually similarity --> see below)
all_delta_VectorNorms_replicates = [] #Dehltave vector norm
all_Gammas_explained_replicates = []

#save the amount of significant concentrations for each drug
num_concentrations = []




#go through all drugs
for d in all_drugs:
    
    #Find the amount of concentrations screened for this drug (sometimes the concentration are not exactly 0.02 but 0.01999)
    used_concentrations = perturbations.loc[perturbations['Drug'] == d]['Concentrations'].unique()
    
    #sort from smallest to largest
    used_concentrations.sort()

    ##find the valid vectors i.e. vectors with norm bigger than previously introduced threshold
    valid_vectors = {}
    valid_vector_norms = {}
    for conc in used_concentrations:
    
        if perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].mean() > threshold_significant:
        
            valid_vectors[conc] = np.array(perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].mean().values)
            valid_vector_norms[conc] = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].mean()
    
            # Make calculations for same replicates
            if len(perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features]) == 2:

                vector1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].values[0]
                vector2 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].values[1]

                vectorNorm1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].values[0]
                vectorNorm2 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].values[1]

                if vectorNorm1 > threshold_significant and vectorNorm2 > threshold_significant:

                    #calculate cosine similarity results
                    cosine_similarity = 1 - spatial.distance.cosine(vector1, vector2)

                    #calculate vector norm result
                    delta_VectorNorm = vectorNorm1 - vectorNorm2      

                    #add results to result lists
                    all_cosine_distances_replicates.append(cosine_similarity)
                    all_delta_VectorNorms_replicates.append(delta_VectorNorm)

                    #perform vector math, use half vectors to not extend to to feaute values that might be not possible (e.g. Eccentricty  0.6 + 0.6 > 1)
                    a,b,g = calculate_vector_math_v2(vector1/2, vector1/2, vector2)

                    #add results to result lists
                    all_AlphaBetas_replicates.append(float(a))
                    all_Gammas_replicates.append(float(g))
                    all_Gammas_explained_replicates.append(float(g)/vectorNorm2)
                
               
                
            

    
    
    #add the number of valid vectors to num_concentration result list
    num_concentrations.append(len(valid_vectors))
    
    # go through the valid concentrations and always pick the consecutive ones
    for i in range(0,len(used_concentrations) - 1):
        
        # take the two concentrations that follow each other
        conc1 = used_concentrations[i] 
        conc2 = used_concentrations[i+1]
        
        #only calculate results if both concentrations are within the valid vectors for this drug
        if conc1 in valid_vectors and conc2 in valid_vectors:
            
            #calculate cosine similarity results
            cosine_similarity = 1 - spatial.distance.cosine(valid_vectors[conc1], valid_vectors[conc2])

            #calculate vector norm result
            delta_VectorNorm = valid_vector_norms[conc2] - valid_vector_norms[conc1]      

            #add results to result lists
            all_cosine_distances.append(cosine_similarity)
            all_delta_VectorNorms.append(delta_VectorNorm)

            
            #print valid_vector_norms[conc1]     
            #print valid_vector_norms[conc2]
            #print valid_vectors[conc1][0:6]
            #print valid_vectors[conc2][0:6]
            #print calculate_vector_math_v2(valid_vectors[conc1]/2, valid_vectors[conc1]/2, valid_vectors[conc2])
            
            
            #perform vector math, use half vectors to not extend to to feaute values that might be not possible (e.g. Eccentricty  0.6 + 0.6 > 1)
            
            #if valid_vector_norms[conc1] < valid_vector_norms[conc2]: 
            a,b,g = calculate_vector_math_v2(valid_vectors[conc1]/2, valid_vectors[conc1]/2, valid_vectors[conc2])
            #else:
            #    a,b,g = calculate_vector_math_v2(valid_vectors[conc2]/2, valid_vectors[conc2]/2, valid_vectors[conc1])

            #add results to result lists
            all_AlphaBetas.append(float(a))
            all_Gammas.append(float(g))
            all_gamma_explained.append(float(g)/valid_vector_norms[conc2])
#Show output
print 'Finished real calculation' 
print 'Number of real entries: %d' %len(all_cosine_distances)

# Show output for significant concentrations per drug
##
print '0 concentrations significant: %d' %num_concentrations.count(0)
print '1 concentrations significant: %d' %num_concentrations.count(1)
print '2 concentrations significant: %d' %num_concentrations.count(2)
print '3 concentrations significant: %d' %num_concentrations.count(3)
print '4 concentrations significant: %d' %num_concentrations.count(4)
print '-----'
print 'More than 1 significant: %d' %len([x for x in num_concentrations if x >1])

# Make bar plot
##

plt.bar([0,1,2,3,4],[num_concentrations.count(0),num_concentrations.count(1),num_concentrations.count(2),num_concentrations.count(3),num_concentrations.count(4)], color='#3AB9D1')
plt.xlabel('Significant Concentrations')
plt.ylabel('Frequency')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Significant_Perturbations_perDrug.pdf')
plt.close()

### Calculate random results
1. Go through all drugs
2. Find significant concentration for the first drug
3. Find all significant consecutive concentrations of all but the first drug
4. Do this for all drugs

# list of concentrations
next_concentrations = {'0.02':'0.2', '0.2':'2.0','2.0':'20.0'}

#Random result lists
all_AlphaBetas_random = []
all_Gammas_random = []
all_cosine_distances_random = []
all_delta_VectorNorms_random = []


all_Gammas_random_explained = []


drug_conc = perturbations.groupby(['Drug','Concentrations'], axis=0, as_index = False).mean()
drug_conc = drug_conc.loc[drug_conc['VectorNorm'] > threshold_significant][['Drug','Concentrations']]


iteration = 0
#go through all rows 
for row in drug_conc.iterrows():
    #select drug and concentrations
    drug1 = row[1][0]
    conc1_orig = row[1][1]
    
    #iteration +=1
    #if iteration > 2:
    #    break
    
    Vector_Length1 = perturbations.loc[(perturbations['Drug'] == drug1) & (perturbations['Concentrations'] == conc1_orig)]['VectorNorm'].mean()
    if Vector_Length1 > threshold_significant:
        vector1 = perturbations.loc[(perturbations['Drug'] == drug1) & (perturbations['Concentrations'] == conc1_orig)][features].mean().values
    
        #print drug1
        #print conc1_orig
        #print Vector_Length1
        #print vector1
        #print '-'
        
         #sometimes concentrations are not exactly on the four points (adjust to range)
        if float(conc1_orig) > 7:
            conc1 = '20.0'
        elif float(conc1_orig) > 0.7:
            conc1 = '2.0'
        elif float(conc1_orig) > 0.07:
            conc1 = '0.2'
        else:
            conc1 = '0.02'

        # if already max concentration go to next row
        if conc1 == '20.0':
            continue

        #find next concentration (consecutive)
        next_concentration = next_concentrations[conc1]
        
        #find all other valid drug/concentrations (must be not the same drug, and the EXACT following concentration)
        other_valid_drugs = drug_conc.loc[(drug_conc['Concentrations'] == next_concentration) & (drug_conc['Drug'] != drug1)]

        #Go through all thre rows of other significant drugs
        for row2 in other_valid_drugs.iterrows():
            #select other drug and concentration
            drug2 = row2[1][0]
            conc2 = row2[1][1]
                                
            Vector_Length2 = perturbations.loc[(perturbations['Drug'] == drug2) & (perturbations['Concentrations'] == conc2)]['VectorNorm'].mean()
            
            if Vector_Length2 > threshold_significant:
                vector2 = perturbations.loc[(perturbations['Drug'] == drug2) & (perturbations['Concentrations'] == conc2)][features].mean().values
                
                #print drug2
                #print conc2
                #print Vector_Length2
                #print vector2
                
                #calculate cosine similarity and delta vector norm
                cosine_similarity = 1 - spatial.distance.cosine(vector1, vector2)
                delta_VectorNorm = Vector_Length2 - Vector_Length1 

                #add results to result lists
                all_cosine_distances_random.append(cosine_similarity)
                all_delta_VectorNorms_random.append(delta_VectorNorm)

                #perform vector math
                #if Vector_Length1 < Vector_Length2:
                a,b,g = calculate_vector_math_v2(vector1/2, vector1/2, vector2)
                #else:
                #    a,b,g = calculate_vector_math_v2(vector2/2, vector2/2, vector1)

                #add results to result lists
                all_AlphaBetas_random.append(float(a))
                all_Gammas_random.append(float(g))
                all_Gammas_random_explained.append(float(g)/Vector_Length2)
                
                #print delta_VectorNorm
                #print g
                #print  calculate_vector_math_v2(vector1/2, vector1/2, vector2)
                #print calculate_vector_math_v2(vector2/2, vector2/2, vector1)
                #print 'xxx'
                
        #print '--'

#Show output
print 'Finished Randomization' 
print 'Number of random entries: %d' %len(all_cosine_distances_random)


## Plot Results

### Cosine similarity and Delta vector norm

print 'Cosine Distance'
print 'Mean replicate results: %.2f' %np.mean(all_cosine_distances_replicates)
print 'Mean real results: %.2f' %np.mean(all_cosine_distances)
print 'Mean random results: %.2f' %np.mean(all_cosine_distances_random)
print mannwhitneyu(all_cosine_distances,all_cosine_distances_random) #Concentration/Random
print mannwhitneyu(all_cosine_distances,all_cosine_distances_replicates) #Concentration/Same
print mannwhitneyu(all_cosine_distances_replicates,all_cosine_distances_random) #Same/Random
print '--\n'
print 'Vector Distance'
print 'Mean replicate results: %.2f' %np.mean(all_delta_VectorNorms_replicates)
print 'Mean real results: %.2f' %np.mean(all_delta_VectorNorms)
print 'Mean random results: %.2f' %np.mean(all_delta_VectorNorms_random)
print mannwhitneyu(all_delta_VectorNorms,all_delta_VectorNorms_random) #Concentration/Random
print mannwhitneyu(all_delta_VectorNorms,all_delta_VectorNorms_replicates) #Concentration/Same
print mannwhitneyu(all_delta_VectorNorms_replicates,all_delta_VectorNorms_random) #replicates/Random

#Make output folder
ensure_dir('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Barplot_CosineSimilarity.pdf')

# Scatter plot Cosine similarityh against Delta vector norm (REAL)
##

plt.scatter(all_delta_VectorNorms,all_cosine_distances, alpha=0.4, color='#3AB9D1')
plt.axvline(0, color='grey', ls='--')
plt.axhline(0, color='grey', ls='--')
plt.xlabel('Delta Vector Norm')
plt.ylabel('Cosine Similarity')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Scatter_RealPoints.pdf')
plt.close()

# Scatter plot Cosine similarityh against Delta vector norm (RANDOM)
##

plt.scatter(all_delta_VectorNorms_random,all_cosine_distances_random, alpha=0.05, color='#3AB9D1')
plt.axvline(0, color='grey', ls='--')
plt.axhline(0, color='grey', ls='--')
plt.xlabel('Delta Vector Norm')
plt.ylabel('Cosine Similarity')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Scatter_RandomPoints.pdf')
plt.close()

# KDE (Density) plot with random results 
# + Scatter plot Cosine similarityh against Delta vector norm (REAL)
# Overlay
##

sns.kdeplot(all_delta_VectorNorms_random, all_cosine_distances_random, shade=True, shade_lowest=False, gridsize=200, bw=0.25, color='grey')
plt.scatter(all_delta_VectorNorms,all_cosine_distances, alpha=0.8, color='#3AB9D1')
plt.axvline(0, color='grey', ls='--')
plt.axhline(0, color='grey', ls='--')
plt.xlabel('Delta Vector Norm')
plt.ylabel('Cosine Similarity')
plt.ylim([-1,1])
plt.xlim([-2,3]) # CHECK THIS before 
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Density_RandomPoints.pdf')
plt.close()

# Bar plot difference in cosine similarity
##


plt.bar([0,1,2],[np.mean(all_cosine_distances_replicates),np.mean(all_cosine_distances), np.mean(all_cosine_distances_random)], 
        yerr = [1.96 * np.std(all_cosine_distances_replicates)/np.sqrt(len(all_cosine_distances_replicates))
            ,1.96 * np.std(all_cosine_distances)/np.sqrt(len(all_cosine_distances))
                ,1.96 * np.std(all_cosine_distances_random)/np.sqrt(len(all_cosine_distances_random))], color='#3AB9D1')
plt.xlabel('Group')
plt.xticks([0,1,2],['Replicates','Real','Random'])
plt.ylabel('Cosine Similarity [95% CI]')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Barplot_CosineSimilarity.pdf')
plt.close()

# Bar plot difference in delta vector norm
##


plt.bar([0,1,2],[np.mean(all_delta_VectorNorms_replicates), np.mean(all_delta_VectorNorms), np.mean(all_delta_VectorNorms_random)],
        yerr = [1.96 * np.std(all_delta_VectorNorms_replicates)/np.sqrt(len(all_delta_VectorNorms_replicates))
            ,1.96 * np.std(all_delta_VectorNorms)/np.sqrt(len(all_delta_VectorNorms))
                ,1.96 * np.std(all_delta_VectorNorms_random)/np.sqrt(len(all_delta_VectorNorms_random))], color='#3AB9D1')
plt.xlabel('Group')
plt.xticks([0,1,2],['Replicates','Real','Random'])
plt.ylabel('Delta Vector Norm [95% CI]')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Barplot_DeltaVectorNorm.pdf')
plt.close()

### Alpha/Beta vector math results

print 'Alpha Beta:'
print mannwhitneyu(all_AlphaBetas_replicates,all_AlphaBetas_random) #Replicate/Random
print mannwhitneyu(all_AlphaBetas,all_AlphaBetas_replicates) #Replicate/Concentration
print mannwhitneyu(all_AlphaBetas,all_AlphaBetas_random) #Concentration/Random
print 'Mean replicate results: %.2f' %np.mean(all_AlphaBetas_replicates)
print 'Mean real results: %.2f' %np.mean(all_AlphaBetas)
print 'Mean random results: %.2f' %np.mean(all_AlphaBetas_random)
print '--\n'
print 'Gamma:'
print mannwhitneyu(all_Gammas_random,all_Gammas_replicates) #Replicate/Random
print mannwhitneyu(all_Gammas,all_Gammas_replicates) #Replicate/Concentration
print mannwhitneyu(all_Gammas,all_Gammas_random) #Concentration/random
print 'Mean replicate results: %.2f' %np.mean(all_Gammas_replicates)
print 'Mean real results: %.2f' %np.mean(all_Gammas)
print 'Mean random results: %.2f' %np.mean(all_Gammas_random)

# KDE (Density) plot with random results 
# + Scatter plot alpha/beta vs gamma (REAL)
# Overlay
##

sns.kdeplot(all_AlphaBetas_random, all_Gammas_random, shade=True, shade_lowest=False, gridsize=200, bw=0.25, color='grey')
plt.scatter(all_AlphaBetas,all_Gammas, alpha=0.8, color='#3AB9D1')
plt.axvline(0, color='grey', ls='--')
plt.axhline(0, color='grey', ls='--')
plt.xlabel('Alpha/Beta')
plt.ylabel('Gamma')
#plt.ylim([-1,1])
#plt.xlim([-2,3]) # CHECK THIS before 
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Density_AlphaBeta.pdf')
plt.close()

# Bar plot difference in alpha/beta contribution
##

plt.bar([0,1,2],[np.mean(all_AlphaBetas_replicates),np.mean(all_AlphaBetas), np.mean(all_AlphaBetas_random)], 
        yerr = [1.96 * np.std(all_AlphaBetas_replicates)/np.sqrt(len(all_AlphaBetas_replicates))
            ,1.96 * np.std(all_AlphaBetas)/np.sqrt(len(all_AlphaBetas))
                ,1.96 * np.std(all_AlphaBetas_random)/np.sqrt(len(all_AlphaBetas_random))], color='#3AB9D1')

plt.xlabel('Group')
plt.xticks([0,1,2],['Replicates','Real','Random'])
plt.ylabel('Alpha/Beta [95% CI]')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Barplot_AlphaBeta.pdf')
plt.close()

# Bar plot difference in gamma contribution
##

plt.bar([0,1,2],[np.mean(all_Gammas_replicates),np.mean(all_Gammas), np.mean(all_Gammas_random)], 
        yerr = [1.96 * np.std(all_Gammas_replicates)/np.sqrt(len(all_Gammas_replicates))
            ,1.96 * np.std(all_Gammas)/np.sqrt(len(all_Gammas))
                ,1.96 * np.std(all_Gammas_random)/np.sqrt(len(all_Gammas_random))], color='#3AB9D1')

plt.xlabel('Group')
plt.xticks([0,1,2],['Replicates','Real','Random'])
plt.ylabel('Gamma [95% CI]')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Barplot_Gamma.pdf')
plt.close()

# Scatter plot alpha/beta agains gamma (REAL)
##

plt.scatter(all_AlphaBetas,all_Gammas, alpha =0.4, color='#3AB9D1')
plt.xlabel('AlphaBeta')
plt.ylabel('Gamma')
plt.axvline(0, color='grey', ls='--')
plt.axhline(np.mean(all_Gammas_random), color='grey', ls='--')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Scatter_AlphaBeta_Gamma.pdf')
plt.close()

# Scatter plot alpha/beta agains gamma (RANDOM)
##

plt.scatter(all_AlphaBetas_random,all_Gammas_random, alpha =0.4, color='#3AB9D1')
plt.xlabel('AlphaBeta')
plt.ylabel('Gamma')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Scatter_AlphaBeta_Gamma_random.pdf')
plt.close()

# Gamma increases with increasing vector norms
# Normalize by subracting the part that is just contributed to increase in vector
##
gamma_normalized_replicates = []
for x,y in zip(all_delta_VectorNorms_replicates,all_Gammas_replicates):
    gamma_normalized_replicates.append(y-x)


gamma_normalized = []
for x,y in zip(all_delta_VectorNorms,all_Gammas):
    gamma_normalized.append(y-x)
    
    
gamma_normalized_random = []
for x,y in zip(all_delta_VectorNorms_random,all_Gammas_random):
    gamma_normalized_random.append(y-x)
    
print 'Gamma_norm:'
print mannwhitneyu(gamma_normalized,gamma_normalized_random)

# Plot gamma normalized differences between real and random
##

plt.bar([0,1,2],[np.mean(gamma_normalized_replicates), np.mean(gamma_normalized), np.mean(gamma_normalized_random)], 
        yerr = [1.96 * np.std(gamma_normalized_replicates)/np.sqrt(len(gamma_normalized_replicates)),
            1.96 * np.std(gamma_normalized)/np.sqrt(len(gamma_normalized)),
                1.96 * np.std(gamma_normalized_random)/np.sqrt(len(gamma_normalized_random))], color='#3AB9D1')

plt.xlabel('Group')
plt.xticks([0,1,2],['Replicates','Real','Random'])
plt.ylabel('Gamma - Delta Vector')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Barplot_Gamma_normalized.pdf')
plt.close()

plt.bar([0,1,2],[np.mean(all_Gammas_explained_replicates), np.mean(all_gamma_explained), np.mean(all_Gammas_random_explained)], 
        yerr = [1.96 * np.std(all_Gammas_explained_replicates)/np.sqrt(len(all_Gammas_explained_replicates)),
            1.96 * np.std(all_gamma_explained)/np.sqrt(len(all_gamma_explained)),
                1.96 * np.std(all_Gammas_random_explained)/np.sqrt(len(all_Gammas_random_explained))], color='#3AB9D1')

plt.xlabel('Group')
plt.xticks([0,1,2],['Replicates','Real','Random'])
plt.ylabel('Gamma/||Vec2||')
plt.savefig('../results/IsabelCLOUPAC_Per_Image/Analyse_Perturbations_over_Concentrations/Barplot_GammaExplained.pdf')
plt.close()

## Reald draws
####


# save results in vectors (consecutive concentrations)
all_AlphaBetas = [] #alpha/beta
all_Gammas = [] #gamma
all_cosine_distances = [] #cosine simialrity (actually similarity --> see below)
all_delta_VectorNorms = [] #Dehltave vector norm


# save results in vectors (consecutive concentrations
all_AlphaBetas_replicates = [] #alpha/beta
all_Gammas_replicates = [] #gamma
all_cosine_distances_replicates = [] #cosine simialrity (actually similarity --> see below)
all_delta_VectorNorms_replicates = [] #Dehltave vector norm


#save the amount of significant concentrations for each drug
num_concentrations = []




#go through all drugs
for d in all_drugs:
    
    #Find the amount of concentrations screened for this drug (sometimes the concentration are not exactly 0.02 but 0.01999)
    used_concentrations = perturbations.loc[perturbations['Drug'] == d]['Concentrations'].unique()
    
    #sort from smallest to largest
    used_concentrations.sort()

    ##find the valid vectors i.e. vectors with norm bigger than previously introduced threshold
    valid_vectors = {}
    valid_vector_norms = {}
    for conc in used_concentrations:
       
        
        
        
       # vector1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].values[0]
        #vector2 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].values[1]

        #vectorNorm1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].values[0]
        #vectorNorm2 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].values[1]
        
        #if vectorNorm1 > threshold_significant or vectorNorm1 > threshold_significant:
        #    if vectorNorm1 > threshold_significant and vectorNorm1 > threshold_significant:
        #        print vector1
        #        print vector2
        #        break
                
                
                #valid_vectors[conc] = 
                #valid_vector_norms[conc] = 
        
        #if perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].mean() > threshold_significant:
        
            #valid_vectors[conc] = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].mean().values
            #valid_vector_norms[conc] = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].mean()
    
        # Make calculations for same replicates
        if len(perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features]) == 2:

            vector1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].values[0]
            vector2 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].values[1]

            vectorNorm1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].values[0]
            vectorNorm2 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].values[1]

            if vectorNorm1 > threshold_significant and vectorNorm2 > threshold_significant:

                #calculate cosine similarity results
                cosine_similarity = 1 - spatial.distance.cosine(vector1, vector2)

                #calculate vector norm result
                delta_VectorNorm = vectorNorm1 - vectorNorm2      

                #add results to result lists
                all_cosine_distances_replicates.append(cosine_similarity)
                all_delta_VectorNorms_replicates.append(delta_VectorNorm)

                #perform vector math, use half vectors to not extend to to feaute values that might be not possible (e.g. Eccentricty  0.6 + 0.6 > 1)
                a,b,g = calculate_vector_math_v2(vector1/2, vector1/2, vector2)

                #add results to result lists
                all_AlphaBetas_replicates.append(float(a))
                all_Gammas_replicates.append(float(g))
                
                
                #add combined vector to valid vectors
                vector_combined = np.mean([vector1,vector2],axis=0)
                norm_combined = np.mean([vectorNorm1,vectorNorm2])
                valid_vectors[conc] = vector_combined
                valid_vector_norms[conc] = norm_combined
            
            elif vectorNorm1 > threshold_significant and vectorNorm2 < threshold_significant:
                valid_vectors[conc] = vector1
                valid_vector_norms[conc] = vectorNorm1
            elif vectorNorm1 < threshold_significant and vectorNorm2 > threshold_significant:
                valid_vectors[conc] = vector2
                valid_vector_norms[conc] = vectorNorm2
        
        elif len(perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features]) == 1:
                vector1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)][features].values[0]
                vectorNorm1 = perturbations.loc[(perturbations['Drug'] == d) & (perturbations['Concentrations'] == conc)]['VectorNorm'].values[0]
                
                if vectorNorm1 > threshold_significant:
                    valid_vectors[conc] = vector1
                    valid_vector_norms[conc] = vectorNorm1
                    
                
            

    
    
    #add the number of valid vectors to num_concentration result list
    num_concentrations.append(len(valid_vectors))
    
    # go through the valid concentrations and always pick the consecutive ones
    for i in range(0,len(used_concentrations) - 1):
        
        # take the two concentrations that follow each other
        conc1 = used_concentrations[i] 
        conc2 = used_concentrations[i+1]
        
        #only calculate results if both concentrations are within the valid vectors for this drug
        if conc1 in valid_vectors and conc2 in valid_vectors:
            
            #calculate cosine similarity results
            cosine_similarity = 1 - spatial.distance.cosine(valid_vectors[conc1], valid_vectors[conc2])
            
            #calculate vector norm result
            delta_VectorNorm = valid_vector_norms[conc2] - valid_vector_norms[conc1]      
            
            #add results to result lists
            all_cosine_distances.append(cosine_similarity)
            all_delta_VectorNorms.append(delta_VectorNorm)
            
            #perform vector math, use half vectors to not extend to to feaute values that might be not possible (e.g. Eccentricty  0.6 + 0.6 > 1)
            a,b,g = calculate_vector_math_v2(valid_vectors[conc1]/2, valid_vectors[conc1]/2, valid_vectors[conc2])
            
            #add results to result lists
            all_AlphaBetas.append(float(a))
            all_Gammas.append(float(g))
#Show output
print 'Finished real calculation' 
print 'Number of real entries: %d' %len(all_cosine_distances)

# list of concentrations
next_concentrations = {'0.02':'0.2', '0.2':'2.0','2.0':'20.0'}

#select all significant perturbations
significant_drug_concentrations = perturbations.loc[perturbations['VectorNorm']  > threshold_significant][['Drug','Concentrations']]


#Random result lists
all_AlphaBetas_random = []
all_Gammas_random = []

all_cosine_distances_random = []
all_delta_VectorNorms_random = []


#go through all rows 
for row in significant_drug_concentrations.iterrows():
    break
    #select drug and concentrations
    drug1 = row[1][0]
    conc1_orig = row[1][1]
    
    #sometimes concentrations are not exactly on the four points (adjust to range)
    if float(conc1_orig) > 7:
        conc1 = '20.0'
    elif float(conc1_orig) > 0.7:
        conc1 = '2.0'
    elif float(conc1_orig) > 0.07:
        conc1 = '0.2'
    else:
        conc1 = '0.02'
    
    # if already max concentration go to next row
    if conc1 == '20.0':
        continue
        
    #find next concentration (consecutive)
    next_concentration = next_concentrations[conc1]

    #find all other valid drug/concentrations (must be not the same drug, and the EXACT following concentration)
    other_valid_drugs = significant_drug_concentrations.loc[(significant_drug_concentrations['Concentrations'] == next_concentration) & (significant_drug_concentrations['Drug'] != drug1)]

    #Go through all thre rows of other significant drugs
    for row2 in other_valid_drugs.iterrows():

        #select other drug and concentration
        drug2 = row2[1][0]
        conc2 = row2[1][1]

        #select vectors
        vector1 = perturbations.loc[(perturbations['Drug'] == drug1) & (perturbations['Concentrations'] == conc1_orig)][features].mean().values
        vector2 = perturbations.loc[(perturbations['Drug'] == drug2) & (perturbations['Concentrations'] == conc2)][features].mean().values

        #select vector norms
        Vector_Length1 = perturbations.loc[(perturbations['Drug'] == drug1) & (perturbations['Concentrations'] == conc1_orig)]['VectorNorm'].mean()
        Vector_Length2 = perturbations.loc[(perturbations['Drug'] == drug2) & (perturbations['Concentrations'] == conc2)]['VectorNorm'].mean()

        
        #if Vector_Length1 > Vector_Length2:    

        #calculate cosine similarity and delta vector norm
        cosine_similarity = 1 - spatial.distance.cosine(vector1, vector2)
        delta_VectorNorm = Vector_Length1 - Vector_Length2

        #add results to result lists
        all_cosine_distances_random.append(cosine_similarity)
        all_delta_VectorNorms_random.append(delta_VectorNorm)

        #perform vector math
        a,b,g = calculate_vector_math_v2(vector1/2, vector1/2, vector2)

        #add results to result lists
        all_AlphaBetas_random.append(float(a))
        all_Gammas_random.append(float(g))

#Show output
print 'Finished Randomization' 
print 'Number of random entries: %d' %len(all_cosine_distances_random)
